SentryMiddleware: Change from `AroundMiddleware` to `Middleware` (#12)
It turns out that `AroundMiddleware` runs much later then a regular `Middleware` independent from the order of `builder.add()` and `build.around()` calls. This migrates the implementation to a regular `Middleware` to avoid this issue and ensure that the request scope is opened as early as possible in the request lifecycle.